# Conflicts:
#	server/build/database/connection.js
#	server/build/database/connection.js.map
#	server/database/connection.ts

+++ frontend/src/app/protected/pages/list-categoria/list-categoria.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { NzModalService } from 'ng-zorro-antd/modal';
+import Swal from 'sweetalert2';
 
 import { Categoria } from '../../interfaces/Categoria';
 import { CategoriasService } from '../../services/categorias.service';
@@ -90,12 +91,33 @@ export class ListCategoriaComponent implements OnInit {
   }
 
   deleteCategoria(id: string){
-    this.categoriaService.deleteCategoria(id).subscribe(
-      res=> {
-        this.getCategorias();
-      },
-      err => console.log(err)
-    );
+
+    Swal.fire({
+      title: 'Esta seguro de eliminar la Categoria?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!',
+      cancelButtonText:'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.categoriaService.deleteCategoria(id).subscribe(
+          res=> {
+            this.getCategorias();
+            Swal.fire(
+              'Eliminado!',
+              'Usted ha eliminado la categoria',
+              'success'
+            )
+          },
+          err => console.log(err)
+        );
+      }
+    })
+
+    
   }
