+++ frontend/src/app/protected/pages/list-medida/list-medida.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 import { UnidadMedidaService } from '../../services/unidad-medida.service';
 
 @Component({
@@ -47,12 +48,35 @@ export class ListMedidaComponent implements OnInit {
   }
 
   deleteMedida(id: string){
-    this.medidaService.deleteMedida(id).subscribe(
-      res=> {
-        this.getMedidas();
-      },
-      err => console.log(err)
-    );
+
+    Swal.fire({
+      title: 'Esta seguro de eliminar el producto?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Eliminar!',
+      cancelButtonText:'Cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.medidaService.deleteMedida(id).subscribe(
+          res=> {
+            this.getMedidas();
+            Swal.fire(
+              'Eliminado!',
+              'Usted ha eliminado la unidad de medida',
+              'success'
+            )
+          },
+          err => console.log(err)
+        );
+
+        
+      }
+    })
+    
+    
   }
 
 }
